test(#7778): cover text block indent through eo-syntax packs instead of a new test class
