+++ eo-parser/src/test/resources/org/eolang/parser/eo-syntax/tab-indented-text-block-body-is-not-under-indented.yaml
@@ -1,0 +1,9 @@
+# SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
+# SPDX-License-Identifier: MIT
+---
+# yamllint disable rule:line-length
+sheets: []
+asserts:
+  - /object[not(errors)]
+  - //o[@base='Φ.string' and @name='txt' and o[@base='Φ.bytes']/o[text()='09-09-66-6F-6F']]
+input: "[] > main\n  \"\"\"\n\t\tfoo\n  \"\"\" > txt"
