+++ eo-parser/src/test/resources/org/eolang/parser/eo-syntax/tab-only-line-in-text-block-is-blank.yaml
@@ -1,0 +1,9 @@
+# SPDX-FileCopyrightText: Copyright (c) 2016-2026 Objectionary.com
+# SPDX-License-Identifier: MIT
+---
+# yamllint disable rule:line-length
+sheets: []
+asserts:
+  - /object[not(errors)]
+  - //o[@base='Φ.string' and @name='txt' and o[@base='Φ.bytes']/o[text()='09-09-09']]
+input: "[] > main\n  \"\"\"\n\t\t\t\n  \"\"\" > txt"
